added ce & pe count and fixed pnl values to 2 digit places
