+++ 07-zerodha-extension/content-utils.js
@@ -1,0 +1,3 @@
+export function test() {
+    console.log('log from content-utils.js')
+}
