+++ 07-zerodha-extension/popup.html
@@ -29,7 +29,6 @@
 <body>
   <b>URL: </b><span id="url"></span> <br><br>
   <button id="add-position-report">Position Report</button> <br><br>
-  <!-- <button id="addButton">Add Watchlist Button</button> -->
 
   <script src="popup.js"></script>
 </body>
