+++ zPOC/countBankExtension/popup.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Bank</title>
+    <style>
+        body {
+            width: 200px;
+            text-align: center;
+            font-family: Arial, sans-serif;
+        }
+        #count {
+            font-size: 24px;
+            font-weight: bold;
+        }
+    </style>
+</head>
+<body>
+    <h2>Word 'Bank' Count</h2>
+    <p id="count">Loading...</p>
+    <script src="popup.js"></script>
+</body>
+</html>
