+++ zPOC/countBankExtension/styles.css
@@ -1,0 +1,11 @@
+.custom-count-box {
+    position: fixed;
+    bottom: 10px;
+    right: 10px;
+    background: rgb(234, 205, 74);
+    color: rgb(46, 20, 20);
+    padding: 10px;
+    border-radius: 5px;
+    font-size: 16px;
+    z-index: 10000;
+}
