+++ zPOC/tempExtension/utils.js
@@ -1,0 +1,3 @@
+export function greet(name) {
+    console.log(`Hello, ${name}!`);
+}
